Refactoring
Move read_config_as_dict, write_config and pretty_print to core
