+++ deeplabcut/pose_estimation_pytorch/config/__init__.py
@@ -16,9 +16,12 @@ from deeplabcut.pose_estimation_pytorch.config.make_pose_config import (
 from deeplabcut.pose_estimation_pytorch.config.utils import (
     available_detectors,
     available_models,
-    pretty_print,
-    read_config_as_dict,
     update_config,
     update_config_by_dotpath,
+)
+# For backwards compatibility
+from deeplabcut.core.config import (
+    read_config_as_dict,
     write_config,
+    pretty_print,
 )
